for #2408, move RuleDefinitionFileConstantTest from sharding-core-parse-test to sharding-core-parse-common
