+++ sharding-core/sharding-core-parse/sharding-core-parse-common/src/test/java/org/apache/shardingsphere/core/parse/rule/AllRuleTests.java
@@ -15,24 +15,14 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.core.parse;
+package org.apache.shardingsphere.core.parse.rule;
 
-import org.apache.shardingsphere.core.parse.constant.AllConstantTests;
-import org.apache.shardingsphere.core.parse.hook.SPIParsingHookTest;
-import org.apache.shardingsphere.core.parse.rule.AllRuleTests;
-import org.apache.shardingsphere.core.parse.sql.context.AllContextTests;
-import org.apache.shardingsphere.core.parse.util.SQLUtilTest;
+import org.apache.shardingsphere.core.parse.rule.jaxb.loader.RuleDefinitionFileConstantTest;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
 
 @RunWith(Suite.class)
-@SuiteClasses({
-        AllConstantTests.class,
-        AllRuleTests.class,
-        SQLUtilTest.class,
-        AllContextTests.class,
-        SPIParsingHookTest.class
-})
-public final class AllParseTests {
+@SuiteClasses(RuleDefinitionFileConstantTest.class)
+public final class AllRuleTests {
 }
